extract go scanner helper to java temp dir

+++ src/main/java/org/pqca/scanning/go/GoScannerService.java
@@ -24,6 +24,7 @@ import jakarta.annotation.Nonnull;
 import jakarta.annotation.Nullable;
 import java.io.File;
 import java.io.IOException;
+import java.nio.file.Files;
 import java.util.ArrayList;
 import java.util.List;
 import org.apache.commons.io.FileUtils;
@@ -38,7 +39,6 @@ import org.pqca.scanning.ScannerService;
 import org.sonar.api.batch.fs.InputFile;
 import org.sonar.api.batch.sensor.SensorContext;
 import org.sonar.api.batch.sensor.internal.SensorContextTester;
-import org.sonar.api.impl.utils.DefaultTempFolder;
 import org.sonar.api.rule.RuleKey;
 import org.sonar.go.converter.GoConverter;
 import org.sonar.go.plugin.GoChecks;
@@ -65,43 +65,52 @@ public final class GoScannerService extends ScannerService {
         int numberOfScannedLines = 0;
         int numberOfScannedFiles = 0;
 
-        File goTempFolder = new DefaultTempFolder(this.projectDirectory).newDir();
-
+        File goTempFolder = null;
         try {
-            GoConverter goConverter = new GoConverter(goTempFolder);
-            GoCheck visitor = new GoDetectionCollectionRule(this);
-            GoChecks checks = new GoRuleChecks(visitor);
-            final SensorContextTester sensorContext = SensorContextTester.create(projectDirectory);
+            goTempFolder = Files.createTempDirectory("cbomkit-go-").toFile();
+        } catch (IOException e) {
+            LOGGER.error("Failed to create temp dir for go scanner", e);
+        }
 
-            // Go scanner (CryptoGoSensor) reads files from context
-            index.forEach(
-                    project -> project.inputFileList().forEach(sensorContext.fileSystem()::add));
-
-            for (ProjectModule project : index) {
-                numberOfScannedFiles += project.inputFileList().size();
-                numberOfScannedLines +=
-                        project.inputFileList().stream().mapToInt(InputFile::lines).sum();
-
-                final String projectStr =
-                        project.identifier() + " (" + counter + "/" + index.size() + ")";
-                if (this.progressDispatcher != null) {
-                    this.progressDispatcher.send(
-                            new ProgressMessage(
-                                    ProgressMessageType.LABEL,
-                                    "Scanning go project " + projectStr));
-                }
-                LOGGER.info("Scanning go project {}", projectStr);
-
-                CryptoGoSensor.execute((SensorContext) sensorContext, goConverter, checks);
-
-                counter += 1;
-            }
-            LOGGER.info("Scanned {} go projects", index.size());
-        } finally {
+        if (goTempFolder != null) {
             try {
-                FileUtils.deleteDirectory(goTempFolder);
-            } catch (IOException e) {
-                LOGGER.error("Failed to delete temp dir {}", goTempFolder);
+                GoConverter goConverter = new GoConverter(goTempFolder);
+                GoCheck visitor = new GoDetectionCollectionRule(this);
+                GoChecks checks = new GoRuleChecks(visitor);
+                final SensorContextTester sensorContext =
+                        SensorContextTester.create(projectDirectory);
+
+                // Go scanner (CryptoGoSensor) reads files from context
+                index.forEach(
+                        project ->
+                                project.inputFileList().forEach(sensorContext.fileSystem()::add));
+
+                for (ProjectModule project : index) {
+                    numberOfScannedFiles += project.inputFileList().size();
+                    numberOfScannedLines +=
+                            project.inputFileList().stream().mapToInt(InputFile::lines).sum();
+
+                    final String projectStr =
+                            project.identifier() + " (" + counter + "/" + index.size() + ")";
+                    if (this.progressDispatcher != null) {
+                        this.progressDispatcher.send(
+                                new ProgressMessage(
+                                        ProgressMessageType.LABEL,
+                                        "Scanning go project " + projectStr));
+                    }
+                    LOGGER.info("Scanning go project {}", projectStr);
+
+                    CryptoGoSensor.execute((SensorContext) sensorContext, goConverter, checks);
+
+                    counter += 1;
+                }
+                LOGGER.info("Scanned {} go projects", index.size());
+            } finally {
+                try {
+                    FileUtils.deleteDirectory(goTempFolder);
+                } catch (IOException e) {
+                    LOGGER.error("Failed to delete temp dir {}", goTempFolder, e);
+                }
             }
         }
